use plain js for simple socks background page
