+++ src/samples/simple-socks-chromeapp/background.js
@@ -1,0 +1,21 @@
+var script = document.createElement('script');
+script.src = 'freedom-for-chrome/freedom-for-chrome.js';
+document.head.appendChild(script);
+
+// Keep a background timeout running continuously, to prevent chrome from
+// putting the app to sleep.
+function keepAlive() {
+  setTimeout(keepAlive, 5000);
+}
+keepAlive();
+
+script.onload = function() {
+  freedom('uproxy-lib/simple-socks/freedom-module.json', {
+    'logger': 'uproxy-lib/loggingprovider/freedom-module.json',
+    'debug': 'debug'
+  }).then(function(moduleFactory) {
+    moduleFactory();
+  }, (e) => {
+    console.error('could not load freedomjs module: ' + e.message);
+  });
+}
